+++ orm/typedef/IPageableResult.ts
@@ -1,0 +1,4 @@
+export default interface IPageableResult<T> {
+    readonly entries: T[];
+    readonly count: string;
+}