+++ orm/typedef/IPaginationOptions.ts
@@ -1,0 +1,4 @@
+export default interface IPaginationOptions {
+    readonly offset: number;
+    readonly limit: number;
+}