+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### BREAKING CHANGES
 
+  -
+
+### Noteworthy
+
+  -
+
+## [4.9.0] - 2021-11-10
+
+### BREAKING CHANGES
+
 -
 
 ### Noteworthy